feat: add model schema and github action to generate translations
